updated Subevent tests and passed all test cases
